+++ tests/CustomAssertions/AnnotationsParser.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 /**
  * This file is part of the bitrix24-php-sdk package.
  *
- * © Maksim Mesilov <mesilov.maxim@gmail.com>
+ * © Gleb Starikov <gleb.starikov1998@mail.ru>
  *
  * For the full copyright and license information, please view the MIT-LICENSE.txt
  * file that was distributed with this source code.