+++ tests/Integration/Services/CRM/Activity/PhantomMethods/ActivityTypePhantomMethods.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 /**
  * This file is part of the bitrix24-php-sdk package.
  *
- * © Maksim Mesilov <mesilov.maxim@gmail.com>
+ * © Gleb Starikov <gleb.starikov1998@mail.ru>
  *
  * For the full copyright and license information, please view the MIT-LICENSE.txt
  * file that was distributed with this source code.
@@ -38,7 +38,7 @@ class ActivityTypePhantomMethods
 
         if ($list !== [] && is_array($list)) {
 
-            $res = $list[0]->getData();
+            $res = $list[0]->getIterator()->getArrayCopy();
 
             $i = 0;
             foreach (array_keys($res) as $key) {
@@ -62,7 +62,7 @@ class ActivityTypePhantomMethods
 
         if ($list !== [] && is_array($list)) {
 
-            $res = $list[0]->getData();
+            $res = $list[0]->getIterator()->getArrayCopy();
 
             $i = 0;
             foreach ($res as $key => $value) {